+++ lib/broker/htmlpages.h
@@ -13,6 +13,7 @@
 #include <Arduino.h>
 #include "htmlstructure.h"
 #include "htmlstyle.h"
+#include "css.h"
 #include "htmlbrokerform.h"
 #include "htmlclientform.h"
 #include "htmltopnav.h"