+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 - Added functionality for irrigation
 - Added switch to enable/disable battery mode
+- Form submit replies with updated form
 
 ## 0.1.0 2020-05-31 Initial revision